requirements.txt
streamlit
faiss-cpu
sentence-transformers
pypdf
python-docx
requests
numpy
rich
